Move SafeUnbuffered to own Python file
